+++ hw12_13_14_15_calendar/internal/server/http/validations.go
@@ -1,0 +1,9 @@
+package http_server
+
+//http.HandleFunc("/get-events-day", applyMiddlewares(getEventsDay, r))
+//http.HandleFunc("/get-events-week", applyMiddlewares(getEventsWeek, r))
+//http.HandleFunc("/get-events-month", applyMiddlewares(getEventsMonth, r))
+//
+//http.HandleFunc("/add-event", applyMiddlewares(addEvent, r))
+//http.HandleFunc("/update-event", applyMiddlewares(updateEvent, r))
+//http.HandleFunc("/delete-event", applyMiddlewares(deleteEvent, r))
